Refactor shared logic into modules
